Update pubspec.lock for package source changes and version updates; add concurrent futures dependency
This commit modifies the pubspec.lock file to change the package source URLs from "https://pub.dev" to "https://pub.flutter-io.cn" for all dependencies, ensuring compatibility with the Chinese Flutter ecosystem. Additionally, it updates the version of the 'background_downloader' package to 9.4.3 and the 'image' package to 4.7.2. Furthermore, a new dependency for 'androidx.concurrent:concurrent-futures-ktx:1.1.0' is added in the build.gradle.kts file to support concurrent futures in the Android application. These changes enhance the application's dependency management and functionality.